pyproject.toml
[build-system]
requires = ["setuptools>=80", "wheel", "setuptools-scm>=9"]
build-backend = "setuptools.build_meta"

[project]
name = "mkzforge"
dynamic = ["version"]
description = "Scripts and tools to ease the processing of videos for Social Media platforms."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Markizano Draconus", email = "markizano@markizano.net" }]

dependencies = [
    "argostranslate>=1.11.0",
    "boto3>=1.42.59",
    "cherrypy>=18.10.0",
    "fabric==3.2.2",
    "google-api-python-client>=2.193.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "google-genai>=1.48.0",
    "jsonschema>=4.25.1",
    "kizano==1.0.6",
    "langchain>=1.0.5",
    "langchain-ollama>=1.0.0",
    "numba>=0.62.1",
    "openai>=2.7.2",
    "openai-whisper==20250625",
    "pillow>=10.0.0",
    "python-ffmpeg==2.0.12",
    "pyyaml>=6.0.1",
    "requests>=2.32.5",
]

[project.scripts]
mkzforge = "mkzforge:main"

[project.urls]
Homepage = "https://mkzforge.markizano.net"
Repository = "https://github.com/markizano/mkzforge"
Issues = "https://github.com/markizano/mkzforge/issues"

[tool.setuptools_scm]
tag_regex = "^v(?P<version>.+)$"
version_scheme = "no-guess-dev"
local_scheme = "node-and-date"
write_to = "lib/mkzforge/_version.py"
write_to_template = "__version__ = \"{version}\"\n"

[tool.setuptools.package-dir]
mkzforge = "lib/mkzforge"

[tool.setuptools.packages.find]
where = ["lib"]

[tool.setuptools.package-data]
mkzforge = [
    "schema.json"
]

[tool.setuptools.data-files]
"share/mkzforge/web" = [
    "web/*.html",
    "web/*.css",
    "web/*.js"
]
